+++ src/main/java/in/hocg/web/global/config/httpsession/Initializer.java
@@ -1,0 +1,10 @@
+package in.hocg.web.global.config.httpsession;
+
+import org.springframework.session.web.context.AbstractHttpSessionApplicationInitializer;
+
+/**
+ * Created by hocgin on 2017/12/24.
+ * email: hocgin@gmail.com
+ */
+public class Initializer extends AbstractHttpSessionApplicationInitializer {
+}